requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
oauthlib==3.1.0
pyOpenSSL==19.0.0
python-dotenv==0.10.3
requests==2.22.0
